requirements.txt
pillow
keras
matplotlib
scikit-learn
pillow
numpy